Disaster-Response-Platform/annoserver/requirements.txt
fastapi==0.95.1
hypothesis-api==1.0.2
pydantic==1.10.13
requests==2.30.0

starlette~=0.26.1